+++ sites/geohub/CHANGELOG.md
@@ -1,5 +1,18 @@
 # geohub
 
+## 1.7.24
+
+### Patch Changes
+
+- b2e0928: fix: removed scroll from vector filter component since the parent component has it. Use unified button design for some buttons of filter.
+
+## 1.7.23
+
+### Patch Changes
+
+- fcbae9f: fix: fixed layout collapse for vector filter
+- 39c3426: fix: to run auth.js correctly, set a constant string if it is not provided
+
 ## 1.7.22
 
 ### Patch Changes